pyproject.toml
[tool.poetry]
name = "healthchecks-decorator"
version = "0.1.0"
description = "Healthchecks Decorator"
authors = ["Daniel del Valle <delvalle.dani@gmail.com>"]
license = "MIT"
readme = "README.rst"
homepage = "https://github.com/danidelvalle/healthchecks-decorator"
repository = "https://github.com/danidelvalle/healthchecks-decorator"
documentation = "https://healthchecks-decorator.readthedocs.io"
classifiers = [
    "Development Status :: 3 - Alpha",
]

[tool.poetry.urls]
Changelog = "https://github.com/danidelvalle/healthchecks-decorator/releases"

[tool.poetry.dependencies]
python = "^3.7"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
coverage = {extras = ["toml"], version = "^6.1"}
safety = "^1.10.3"
mypy = "^0.910"
typeguard = "^2.13.2"
xdoctest = {extras = ["colors"], version = "^0.15.10"}
sphinx = "^4.3.0"
sphinx-autobuild = ">=2021.3.14"
pre-commit = "^2.15.0"
flake8 = "^4.0.1"
black = ">=21.10b0"
flake8-bandit = "^2.1.2"
flake8-bugbear = "^21.9.2"
flake8-docstrings = "^1.6.0"
flake8-rst-docstrings = "^0.2.3"
pep8-naming = "^0.12.1"
darglint = "^1.8.1"
reorder-python-imports = "^2.6.0"
pre-commit-hooks = "^4.0.1"
sphinx-click = "^3.0.2"
Pygments = "^2.10.0"
pyupgrade = "^2.29.1"
furo = ">=2021.11.12"

[tool.poetry.scripts]
healthchecks-decorator = "healthchecks_decorator.__main__:main"

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
source = ["healthchecks_decorator", "tests"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
